fix: 修正cache_ttl配置项的类型定义

+++ src/nonebot_plugin_cs2match/config.py
@@ -16,5 +16,5 @@ class Config(BaseModel):
         ("cac", 60, ["cac"]),
         ("pnl", 50, ["pnl"]),
     ]  # 赛事系列，优先级，匹配赛事名称（小写）
-    cache_ttl: int = 60
+    cache_ttl: float = 60
     cache_max_size: int = 64
